move radio button
