+++ app/src/atoms/buttons/index.ts
@@ -3,7 +3,6 @@ export { FloatingActionButton } from './FloatingActionButton'
 export { LargeButton } from './LargeButton'
 export { MediumButton } from './MediumButton'
 export { QuaternaryButton } from './QuaternaryButton'
-export { RadioButton } from './RadioButton'
 export { SmallButton } from './SmallButton'
 export { SubmitPrimaryButton } from './SubmitPrimaryButton'
 export { TertiaryButton } from './TertiaryButton'
