+++ components/src/atoms/buttons/index.ts
@@ -1,3 +1,4 @@
 export * from './AlertPrimaryButton'
 export * from './PrimaryButton'
 export * from './SecondaryButton'
+export * from './RadioButton'
